feat(python): resolve mastering delivery targets and align CLI options
- Add mastering_platform_names() and resolve targetPlatform through the library, so the assistant takes a delivery-target name rather than a numeric index
- Report loudness_target_limited on the stereo mastering result and mirror the new C struct field in ctypes
- Add ErrorCode.ENCODE_FAILED and its CLI exit status
- Declare _c_aliases_ on the mirrors reused for the engine-side struct typedefs so the layout snapshot checks each C struct they are passed to, and correct three project signatures to c_size_t
- Compare every declared ctypes argtypes and restype positionally against the signature in the C headers by ABI shape rather than by ctypes class, so interchangeable spellings agree while a width or sign change fails, with floors on the scanned declarations and recorded signatures so the comparison cannot pass by comparing nothing
- Accept --detector-delay-ms with --lookahead-ms as its former spelling, give rhythm, dynamics, melody and pcen the DSP options they read, and make the mixing-preset default one the command can run
- Print the diagnostics of project validate --strict in place of the success line, and send advisory diagnostics to stderr
- Treat EngineMidiClipSchedule.destination_id as None for "follow the track" so an explicit 0 is forwarded unchanged
- Prefer the development build directory over the package-adjacent library so a source checkout cannot load a stale copy
- Clamp a non-finite sample the way the C++ WAV writer does instead of raising from int(round(...))
- Reject 0 for the decompose_stems count arguments rather than reading it as the C ABI's default sentinel, and document it
- Compare the stub's TypedDict keys against the runtime declarations, and generate the SoloProcessor literal from the library

+++ bindings/python/src/libsonare/_ffi_signatures_project.py
@@ -44,7 +44,7 @@ def configure_project_signatures(lib: ctypes.CDLL) -> None:
         lib.sonare_project_unresolved_audio_source_id_by_index.restype = ctypes.c_int32
         lib.sonare_project_unresolved_audio_source_id_by_index.argtypes = [
             ctypes.c_void_p,
-            ctypes.c_uint32,
+            ctypes.c_size_t,
             ctypes.POINTER(ctypes.c_uint32),
         ]
         lib.sonare_project_set_source_audio.restype = ctypes.c_int32
@@ -257,7 +257,7 @@ def configure_project_signatures(lib: ctypes.CDLL) -> None:
             ctypes.c_void_p,
             ctypes.c_uint32,
             ctypes.POINTER(SonareProjectClipTake),
-            ctypes.c_uint32,
+            ctypes.c_size_t,
             ctypes.c_uint32,
         ]
 
@@ -266,7 +266,7 @@ def configure_project_signatures(lib: ctypes.CDLL) -> None:
             ctypes.c_void_p,
             ctypes.c_uint32,
             ctypes.POINTER(SonareProjectClipCompSegment),
-            ctypes.c_uint32,
+            ctypes.c_size_t,
         ]
 
         lib.sonare_project_set_warp_map.restype = ctypes.c_int32
